changed so fma timer will be slightly earlier than normal after bind clicks
changed so binds reset coundowns to the bottom of every second. the fma countdowns right now are reset to the top of the second every bind. multiple binds make fma much later than it seems with the addition of origin binds. if i eventually make the timers more precise, i will remove this change